Machine Gunner (#35)
* Second_Robot

* MachineGunner Prefub-Variant

Also put the Turn System at the end of SymulationSystemGroup (so there is an update of visions before the turning off systems)
Corrected AnimDb as well

* Implemented Reload anim + Improved the Logics of Anims

* Deploy and Undeploy Anims Implementation

Also fixed the wrong naming of Deply and Undeploy anims of the MachineGunner

* Fixed Mesh Borders

Also increased the size of MachineGunner

---------

+++ FangsAndSteel/Assets/Scripts/Units/Attack/AttackAuthoring.cs
@@ -9,7 +9,7 @@ public class AttackAuthoring : MonoBehaviour
     public int damage = 0;
     public float realodLen = 1;
     public int attackRadius = 0;
-    public int timeToShoot = 1;
+    public float timeToShoot = 1;
     public class Baker : Baker<AttackAuthoring>
     {
         public override void Bake(AttackAuthoring authoring)
